Moved customizable settings into commands.sh

+++ commands.sh
@@ -1,16 +1,25 @@
 #!/bin/bash
 # Edit your commands in this file.
 
-if ! tmux ls | grep -v send | grep -q $copname; then
-	[ ! -z ${URLS[*]} ] && {
+if [ "$1" = "source" ];then
+	# Edit the token in here
+	TOKEN='tokenhere'
+	# Set INLINE to 1 in order to receive inline queries. 
+	# To enable this option in your bot, send the /setinline command to @BotFather.
+	INLINE=0
+	# Set to .* to allow sending files from all locations
+	FILE_REGEX='/home/user/allowed/.*'
+else
+	if ! tmux ls | grep -v send | grep -q $copname; then
+		[ ! -z ${URLS[*]} ] && {
 		curl -s ${URLS[*]} -o $NAME
-		send_file "${USER[ID]}" "$NAME" "$CAPTION"
-		rm "$NAME"
-	}
-	[ ! -z ${LOCATION[*]} ] && send_location "${USER[ID]}" "${LOCATION[LATITUDE]}" "${LOCATION[LONGITUDE]}"
+			send_file "${USER[ID]}" "$NAME" "$CAPTION"
+			rm "$NAME"
+		}
+		[ ! -z ${LOCATION[*]} ] && send_location "${USER[ID]}" "${LOCATION[LATITUDE]}" "${LOCATION[LONGITUDE]}"
 		
-	# Inline 
-	if [ $INLINE == 1 ]; then
+		# Inline 
+		if [ $INLINE == 1 ]; then
 			# inline query data
 			iUSER[FIRST_NAME]=$(echo "$res" | sed 's/^.*\(first_name.*\)/\1/g' | cut -d '"' -f3 | tail -1)
 			iUSER[LAST_NAME]=$(echo "$res" | sed 's/^.*\(last_name.*\)/\1/g' | cut -d '"' -f3)
@@ -34,8 +43,8 @@ if ! tmux ls | grep -v send | grep -q $copname; then
 				answer_inline_query "$iQUERY_ID" "article" "Telegram" "https://telegram.org/"
 			fi
 		fi
-fi
-case $MESSAGE in
+	fi
+	case $MESSAGE in
 		'/question')
 			startproc "./question"
 			;;
@@ -62,4 +71,5 @@ Contribute to the project: https://github.com/topkecleon/telegram-bot-bash
 		*)
 			if tmux ls | grep -v send | grep -q $copname;then inproc; else send_message "${USER[ID]}" "$MESSAGE" "safe";fi
 			;;
-esac
+	esac
+fi
